Migrate parsers [WIP]
